backend/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.34.0
groq==0.4.2
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-jaeger==1.21.0
python-dotenv==1.0.0
httpx==0.25.2
